fix: Rename file to match file path
The armplugin1.cpp file must be changed to armplugin.cpp for the robot to use the DQN agent in the gazebo simulation. There are two scenarios for the robot, challenge 1 and challenge 2. Armplugin.cpp addresses challenge 1 and armplugin2.cpp addresses challenge 2. The folder is now set to challenge 1. If challenge 2 is requested, please change the filename of armplugin2.cpp to armplugin.cpp and the challenge 1 armplugin.cpp to armplugin1.cpp